+++ src/mocks/data/applications-data.ts
