+++ src/mocks/handlers/bookmarked-recruitments-handler.ts
@@ -1,6 +1,6 @@
 import { API_BASE_URL } from '@/constants/api-constants'
 import { http, HttpResponse } from 'msw'
-import { bookmarkedRecruitmentsMock } from '@/mocks/data/bookmarked-recruitments'
+import { bookmarkedRecruitmentsMock } from '@/mocks/data/bookmarked-recruitments-data'
 
 const getBookmarkedRecruitments = http.get(
   `${API_BASE_URL}/recruitments/bookmarks/me`,
