+++ src/mocks/handlers/chat-handler.ts
@@ -1,6 +1,6 @@
 import { API_BASE_URL } from '@/constants/api-constants'
 import { http, HttpResponse, ws } from 'msw'
-import { chatRoomMessages } from '@/mocks/data/chat-room-messages'
+import { chatRoomMessages } from '@/mocks/data/chat-room-data'
 
 const getChatMessages = http.get<{ study_group_uuid: string }>(
   `${API_BASE_URL}/chat/rooms/:study_group_uuid/messages`,
