+++ src/mocks/handlers/notification-handler.ts
@@ -1,5 +1,5 @@
 import { http, HttpResponse } from 'msw'
-import { mockNotificationList } from '@/mocks/data/notification'
+import { mockNotificationList } from '@/mocks/data/notification-data'
 import { API_BASE_URL } from '@/constants/api-constants'
 
 const getNotifications = http.get(`${API_BASE_URL}/notifications`, () => {
